feat: create one factory

+++ src/controllers/link/deleteLink.ts
@@ -32,7 +32,7 @@ export const deleteLink = async (req: Request, { res }: Response) => {
 export const deleteLinks = async (req: Request, { res }: Response) => {
   try {
     const ids: any[] = req.body;
-    validateIds(ids, res);
+    if (!validateIds(ids, res)) return;
 
     const deletedLink = await Link.deleteMany({ _id: { $in: ids } });
